[DSTEW-1752] - Code review changes - moved error sort to DataClaimsExceptionHandler - simplified error handling in ClaimVersionValidationStep
