+++ claims-data/service/src/main/java/uk/gov/justice/laa/dstew/payments/claimsdata/dto/amendment/ClaimAmendmentValidationCode.java
@@ -30,17 +30,6 @@ public enum ClaimAmendmentValidationCode {
   INVALID_NULL_VERSION(
       ValidationSeverity.FATAL, HttpStatus.BAD_REQUEST, "Claim Version is null", null),
 
-  /** The claim has a null State so cannot be amended. */
-  INVALID_NULL_STATE(ValidationSeverity.FATAL, HttpStatus.BAD_REQUEST, "Claim State is null", null),
-
-  /** The claim has a null Before State so cannot be amended. */
-  INVALID_NULL_BEFORE_STATE(
-      ValidationSeverity.FATAL, HttpStatus.BAD_REQUEST, "Claim Before State is null", null),
-
-  /** The claim has a null Before State so cannot be amended. */
-  INVALID_NULL_PAYLOAD(
-      ValidationSeverity.FATAL, HttpStatus.BAD_REQUEST, "Claim Amendment Payload is null", null),
-
   /** The claim has a stale version number so cannot be amended. */
   INVALID_CLAIM_VERSION_CONFLICT(
       ValidationSeverity.FATAL, HttpStatus.CONFLICT, "Claim Version conflict exists", null),
