+++ claims-data/service/src/test/java/uk/gov/justice/laa/dstew/payments/claimsdata/exception/ClaimAmendmentValidationExceptionTest.java
@@ -13,58 +13,6 @@ import uk.gov.justice.laa.dstew.payments.claimsdata.dto.amendment.ClaimAmendment
 class ClaimAmendmentValidationExceptionTest {
 
   @Test
-  @DisplayName(
-      "Constructor automatically sorts errors by fatality first, then by HTTP status code descending")
-  void shouldSortErrorsByFatalityThenHttpStatusDescending() {
-    // Arrange: Create a mixed pool of fatal and non-fatal errors with various HTTP statuses
-
-    // 1. Non-fatal errors
-    ClaimAmendmentValidationError nonFatal400 = mock(ClaimAmendmentValidationError.class);
-    when(nonFatal400.isFatal()).thenReturn(false);
-    when(nonFatal400.getHttpStatus()).thenReturn(HttpStatus.BAD_REQUEST); // 400
-
-    ClaimAmendmentValidationError nonFatal500 = mock(ClaimAmendmentValidationError.class);
-    when(nonFatal500.isFatal()).thenReturn(false);
-    when(nonFatal500.getHttpStatus()).thenReturn(HttpStatus.INTERNAL_SERVER_ERROR); // 500
-
-    // 2. Fatal errors
-    ClaimAmendmentValidationError fatal400 = mock(ClaimAmendmentValidationError.class);
-    when(fatal400.isFatal()).thenReturn(true);
-    when(fatal400.getHttpStatus()).thenReturn(HttpStatus.BAD_REQUEST); // 400
-
-    ClaimAmendmentValidationError fatal409 = mock(ClaimAmendmentValidationError.class);
-    when(fatal409.isFatal()).thenReturn(true);
-    when(fatal409.getHttpStatus()).thenReturn(HttpStatus.CONFLICT); // 409
-
-    ClaimAmendmentValidationError fatal500 = mock(ClaimAmendmentValidationError.class);
-    when(fatal500.isFatal()).thenReturn(true);
-    when(fatal500.getHttpStatus()).thenReturn(HttpStatus.INTERNAL_SERVER_ERROR); // 500
-
-    // Supply them completely out of order to the constructor
-    List<ClaimAmendmentValidationError> unsortedErrors =
-        List.of(nonFatal400, fatal400, nonFatal500, fatal500, fatal409);
-
-    // Act
-    ClaimAmendmentValidationException exception =
-        new ClaimAmendmentValidationException(unsortedErrors);
-
-    // Assert: Verify strict priority ordering constraints are met
-    List<ClaimAmendmentValidationError> sortedResult = exception.getErrors();
-
-    assertThat(sortedResult).hasSize(5);
-
-    // Priority 1: Fatal errors take precedence, sorted highest HTTP status code first (500 -> 409
-    // -> 400)
-    assertThat(sortedResult.get(0)).isSameAs(fatal500);
-    assertThat(sortedResult.get(1)).isSameAs(fatal409);
-    assertThat(sortedResult.get(2)).isSameAs(fatal400);
-
-    // Priority 2: Non-fatal errors follow, sorted highest HTTP status code first (500 -> 400)
-    assertThat(sortedResult.get(3)).isSameAs(nonFatal500);
-    assertThat(sortedResult.get(4)).isSameAs(nonFatal400);
-  }
-
-  @Test
   @DisplayName("Constructor sets standard RuntimeException failure message with total error count")
   void shouldSetCorrectExceptionMessage() {
     // Arrange
