+++ claims-data/service/src/test/java/uk/gov/justice/laa/dstew/payments/claimsdata/service/amendment/validation/ClaimVersionValidationStepTest.java
@@ -30,7 +30,8 @@ class ClaimVersionValidationStepTest {
 
     // Assert
     assertThat(errors).hasSize(1);
-    assertThat(errors.get(0).getCode()).isEqualTo(ClaimAmendmentValidationCode.INVALID_NULL_STATE);
+    assertThat(errors.get(0).getCode())
+        .isEqualTo(ClaimAmendmentValidationCode.INVALID_NULL_VERSION);
   }
 
   @Test
@@ -46,7 +47,7 @@ class ClaimVersionValidationStepTest {
     // Assert
     assertThat(errors).hasSize(1);
     assertThat(errors.get(0).getCode())
-        .isEqualTo(ClaimAmendmentValidationCode.INVALID_NULL_BEFORE_STATE);
+        .isEqualTo(ClaimAmendmentValidationCode.INVALID_NULL_VERSION);
   }
 
   @Test
@@ -62,7 +63,7 @@ class ClaimVersionValidationStepTest {
     // Assert
     assertThat(errors).hasSize(1);
     assertThat(errors.get(0).getCode())
-        .isEqualTo(ClaimAmendmentValidationCode.INVALID_NULL_PAYLOAD);
+        .isEqualTo(ClaimAmendmentValidationCode.INVALID_NULL_VERSION);
   }
 
   @Test
